of: dynamic: Add of_create_node() and of_destroy_node()
of_create_node() creates device node and apply to base tree dynamically.
The parent device node and full name are required for creating the node.
And the caller can also provide a property array for the node.

Inside this function, it creates a changeset. Then the new device node
and properties are added to the changeset and applied to base tree. The
pointer of this changeset is saved in device node private data.

of_destroy_node() removes the node created by of_create_node() from the
base tree and free it. It gets the changeset pointer from device node
private data and call of_changeset_destroy() to free everything.

Signed-off-by: Lizhi Hou <lizhi.hou@amd.com>
Signed-off-by: Sonal Santan <sonal.santan@amd.com>
Signed-off-by: Max Zhen <max.zhen@amd.com>
Signed-off-by: Brian Xu <brian.xu@amd.com>